requirements.txt
alembic==1.13.2
cffi==1.17.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
Flask==2.2.5
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.2
Mako==1.3.5
MarkupSafe==2.1.5
pycparser==2.22
PyMySQL==1.1.1
pyodbc==5.1.0
pypyodbc==1.3.6
python-dotenv==1.0.1
setuptools==73.0.1
SQLAlchemy==2.0.32
typing_extensions==4.12.2
Werkzeug==2.3.7
WTForms==3.1.2
